Performance getNearestWaypoints
This will speed up the function getNearestWaypoints by 5x.

Idea is to catch up the waypoints in a distance of 50m around the looking position with simple distance function instead of performing complex calculations on all waypoints.
After that the complex calculations are performed on the catched waypoints to find the suitable nearest waypoint.
In case less then 10 waypoints are found the existing approach is followed.

Possible further improvements:
- consider workwith instead of fixed 50m
- consider number of workers(multitools) on same course

Remark:
I will not invest more time on this, so leave it up to you:
Take it, improve it or drop it.